+++ bilrost-derive/src/field/oneof.rs
@@ -1,9 +1,9 @@
 use anyhow::{bail, Error};
 use proc_macro2::TokenStream;
 use quote::quote;
-use syn::punctuated::Punctuated;
-use syn::{Expr, ExprLit, Lit, LitInt, Meta, MetaNameValue, Token, Type};
+use syn::{Meta, Type};
 
+use crate::attrs::tag_list_attr;
 use crate::field::set_option;
 
 #[derive(Clone)]
@@ -18,29 +18,7 @@ impl Field {
         let mut unknown_attrs = Vec::new();
 
         for attr in attrs {
-            if attr.path().is_ident("oneof") {
-                let tags = match attr {
-                    // oneof(1, 2, 3, 4, 5)
-                    Meta::List(meta_list) => meta_list
-                        .parse_args_with(Punctuated::<LitInt, Token![,]>::parse_terminated)?
-                        .iter()
-                        .map(LitInt::base10_parse)
-                        .collect::<Result<Vec<_>, _>>()
-                        .map_err(Error::from),
-                    // oneof = "1, 2, 3, 4, 5"
-                    Meta::NameValue(MetaNameValue {
-                        value:
-                            Expr::Lit(ExprLit {
-                                lit: Lit::Str(lit), ..
-                            }),
-                        ..
-                    }) => lit
-                        .value()
-                        .split(',')
-                        .map(|s| s.trim().parse::<u32>().map_err(Error::from))
-                        .collect::<Result<Vec<u32>, _>>(),
-                    _ => bail!("invalid oneof attribute: {}", quote!(#attr)),
-                }?;
+            if let Some(tags) = tag_list_attr("oneof", Some(100), attr)? {
                 set_option(&mut oneof_tags, tags, "duplicate oneof attribute")?;
             } else {
                 unknown_attrs.push(attr);
@@ -60,7 +38,7 @@ impl Field {
 
         Ok(Some(Field {
             ty: ty.clone(),
-            tags,
+            tags: tags.iter_tags().collect(),
         }))
     }